server/requirements.txt
fastapi
uvicorn
google-genai
azure-cognitiveservices-speech

openai
pyyaml
pydantic
jinja2
python-multipart
mutagen
